+++ components/shared/shell/NavigationItems.tsx
@@ -1,12 +1,4 @@
 import Link from 'next/link';
-import {
-  Cog6ToothIcon,
-  CodeBracketIcon,
-  LockClosedIcon,
-  RectangleStackIcon,
-  UserCircleIcon,
-} from '@heroicons/react/24/outline';
-import { useTranslation } from 'next-i18next';
 import classNames from 'classnames';
 
 export interface MenuItem {
@@ -52,61 +44,6 @@ const NavigationItems = ({ menus }: NavigationItemsProps) => {
       ))}
     </ul>
   );
-};
-
-// User navigation
-export const UserNavigation = ({ activePathname }: NavigationProps) => {
-  const { t } = useTranslation('common');
-
-  const menus: MenuItem[] = [
-    {
-      name: t('all-teams'),
-      href: '/teams',
-      icon: RectangleStackIcon,
-      active: activePathname === '/teams',
-    },
-    {
-      name: t('account'),
-      href: '/settings/account',
-      icon: UserCircleIcon,
-      active: activePathname === '/settings/account',
-    },
-    {
-      name: t('password'),
-      href: '/settings/password',
-      icon: LockClosedIcon,
-      active: activePathname === '/settings/password',
-    },
-  ];
-
-  return <NavigationItems menus={menus} />;
-};
-
-// Team navigation
-export const TeamNavigation = ({
-  slug,
-  activePathname,
-}: NavigationProps & { slug: string }) => {
-  const { t } = useTranslation('common');
-
-  const menus: MenuItem[] = [
-    {
-      name: t('all-products'),
-      href: `/teams/${slug}/products`,
-      icon: CodeBracketIcon,
-      active: activePathname === `/teams/${slug}/products`,
-    },
-    {
-      name: t('settings'),
-      href: `/teams/${slug}/settings`,
-      icon: Cog6ToothIcon,
-      active:
-        activePathname?.startsWith(`/teams/${slug}`) &&
-        !activePathname.includes('products'),
-    },
-  ];
-
-  return <NavigationItems menus={menus} />;
 };
 
 export default NavigationItems;
